+++ src/date-picker/picker.component.ts
@@ -60,6 +60,7 @@ export class ThyPicker implements OnChanges, AfterViewInit {
     @Input() mode: string;
     @Input({ transform: coerceBooleanProperty }) hasBackdrop: boolean;
     @Input() separator: string;
+    @Input() timeZone: string;
     @Output() blur = new EventEmitter<Event>();
     @Output() readonly valueChange = new EventEmitter<TinyDate | TinyDate[] | null>();
     @Output() readonly openChange = new EventEmitter<boolean>(); // Emitted when overlay's open state change
@@ -136,6 +137,9 @@ export class ThyPicker implements OnChanges, AfterViewInit {
                 this.closeDatePopup();
             }
         }
+        if (changes.timeZone && changes.timeZone.currentValue) {
+            this.formatDate(this.innerValue as TinyDate);
+        }
     }
 
     ngAfterViewInit(): void {
@@ -167,7 +171,7 @@ export class ThyPicker implements OnChanges, AfterViewInit {
         if (this.readonlyState) {
             return;
         }
-        this.valueChange.emit(this.pickerInput.nativeElement.value || this.getReadableValue(new TinyDate()));
+        this.valueChange.emit(this.pickerInput.nativeElement.value || this.getReadableValue(new TinyDate(undefined, this.timeZone)));
         this.entering = false;
     }
 
@@ -280,7 +284,7 @@ export class ThyPicker implements OnChanges, AfterViewInit {
         if (this.innerFormat && (this.innerFormat.includes('q') || this.innerFormat.includes('Q'))) {
             return value.format(this.innerFormat);
         } else {
-            return this.dateHelper.format(value.nativeDate, this.innerFormat);
+            return this.dateHelper.format(value?.nativeDate, this.innerFormat);
         }
     }